[ADD] reestructurado y reorganizado reportes l10n_ec
