+++ l10n_ec_reports_penta/report/__init__.py
@@ -3,6 +3,4 @@
 # © 2025 PentaLab
 # License Odoo Proprietary License v1.0 (https://www.odoo.com/documentation/user/16.0/legal/licenses/licenses.html#odoo-proprietary-license)
 
-from . import reports_report
 from . import invoice_report_xlsx
-from . import report_ventas_a1
