+++ l10n_ec_reports_penta/views/menuitems.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <!-- ANEXOS DE IMPUESTOS -->
+    <menuitem id="menu_reports_tax_annexes" name="Anexos de Impuestos" parent="account.menu_finance_reports" sequence="96"/>
+    <menuitem id="menu_action_report_sales_a1" name="A1 Ventas" parent="menu_reports_tax_annexes" action="action_report_sales_a1_wizard" sequence="1" />
+    <menuitem id="menu_action_report_sales_a2" name="A2 Compras" parent="menu_reports_tax_annexes" action="" sequence="2" />
+    <menuitem id="menu_action_report_sales_a3" name="A3 Retenciones compras" parent="menu_reports_tax_annexes" action="" sequence="3" />
+    <menuitem id="menu_action_report_sales_retention" name="Retenciones Ventas" parent="menu_reports_tax_annexes" action="" sequence="4" />
+    <!-- REPORTES PENTA -->
+    <menuitem id="menu_reports_penta" name="Reportes Penta" parent="account.menu_finance_reports" sequence="97"/>
+    <menuitem id="menu_cobros_por_ventas" name="Cobros por Ventas" parent="menu_reports_penta" action="action_cobros_por_ventas" sequence="1"/>
+    <menuitem id="menu_pagos_por_compras" name="Pagos por Compras" parent="menu_reports_penta"
+              action="action_pagos_por_compras" sequence="2"/>
+    <menuitem id="menu_pentalab_report_antiguedad_wizard" name="Antigüedad de cartera V." parent="menu_reports_penta"
+              action="action_pentalab_report_antiguedad_wizard" sequence="3"/>
+    <menuitem id="menu_pentalab_purchase_retentions_wizard" name="Compras y Retenciones" parent="menu_reports_penta"
+              action="action_report_purchase_retentions_wizard_wizard" sequence="4"/>
+    <!-- OTROS REPORTES -->
+    <menuitem id="menu_reports_others" name="Otros Reportes" parent="account.menu_finance_reports" sequence="98"/>
+    <menuitem id="menu_report_uafe" name="UAFE" parent="menu_reports_others" action="action_report_uafe_wizard" sequence="1" />
+
+
+    <menuitem id="menu_reports_report" name="Reports" parent="account.menu_finance_reports" sequence="6" />
+    <menuitem id="menu_action_reports_report" name="Generate reports" parent="menu_reports_report" action="action_generate_reports_wizard" sequence="1" />
+</odoo>
