+++ l10n_ec_reports_penta/wizard/__init__.py
@@ -1,3 +1,6 @@
-from . import generate_reports
 from . import invoice_report_wizard
 from . import pentalab_report_antiguedad_wizard
+
+from . import report_purchase_retentions_wizard
+from . import report_sales_a1_wizard
+from . import report_uafe_wizard
