--- l10n_ec_reports_penta/wizard/generate_reports.xml
@@ -1,61 +1,0 @@
-<odoo>
-
-    <record id="view_generate_reports_wizard_form" model="ir.ui.view">
-        <field name="name">generate.reports.wizard.form</field>
-        <field name="model">generate.reports.wizard</field>
-        <field name="arch" type="xml">
-            <form string="Report Generator">
-                <group>
-                    <field name="report_type"/>
-                    <field name="date_start" required="report_type == 'purchase'" invisible="report_type != 'purchase'"/>
-                    <field name="date_end" required="report_type == 'purchase'" invisible="report_type != 'purchase'"/>
-                    <field name="year" required="report_type == 'uafe'" invisible="report_type != 'uafe'"/>
-                    <field name="month" required="report_type == 'uafe'" invisible="report_type != 'uafe'"/>
-                    <field name="domain_uafe" required="report_type == 'uafe'" invisible="report_type != 'uafe'"/>
-                </group>
-                <footer>
-                    <button name="print_report" type="object" string="Generate" class="btn-primary"/>
-                    <button string="Cancel" special="cancel" class="btn-secondary"/>
-                </footer>
-            </form>
-        </field>
-    </record>
-
-    <record id="action_generate_reports_wizard" model="ir.actions.act_window">
-        <field name="name">Generate Reports</field>
-        <field name="res_model">generate.reports.wizard</field>
-        <field name="view_mode">form</field>
-        <field name="target">new</field>
-    </record>
-
-    <menuitem id="menu_reports_report" name="Reports" parent="account.menu_finance_reports" sequence="6" />
-    <menuitem id="menu_action_reports_report" name="Generate reports" parent="menu_reports_report" action="action_generate_reports_wizard" sequence="1" />
-
-    <record id="view_report_a1_wizard_form" model="ir.ui.view">
-        <field name="name">report.a1.wizard.form</field>
-        <field name="model">report.a1.wizard</field>
-        <field name="arch" type="xml">
-            <form string="Reporte Ventas">
-                <group>
-                    <field name="date_start"/>
-                    <field name="date_end"/>
-                    <field name="document_type"/>
-                </group>
-                <footer>
-                    <button name="print_report" type="object" string="Generate" class="btn-primary"/>
-                    <button string="Cancel" special="cancel" class="btn-secondary"/>
-                </footer>
-            </form>
-        </field>
-    </record>
-
-    <record id="action_report_a1_wizard" model="ir.actions.act_window">
-        <field name="name">Reporte Ventas</field>
-        <field name="res_model">report.a1.wizard</field>
-        <field name="view_mode">form</field>
-        <field name="target">new</field>
-    </record>
-
-    <menuitem id="menu_action_report_ventas" name="Reporte Ventas" parent="menu_reports_report" action="action_report_a1_wizard" sequence="2" />
-
-</odoo>
