+++ l10n_ec_reports_penta/wizard/report_purchase_retentions_wizard_views.xml
@@ -1,0 +1,27 @@
+<odoo>
+
+    <record id="view_report_purchase_retentions_wizard_form" model="ir.ui.view">
+        <field name="name">report.purchase.retentions.wizard.form</field>
+        <field name="model">report.purchase.retentions.wizard</field>
+        <field name="arch" type="xml">
+            <form string="Report Generator">
+                <group>
+                    <field name="date_start"/>
+                    <field name="date_end"/>
+                </group>
+                <footer>
+                    <button name="print_report" type="object" string="Generate" class="btn-primary"/>
+                    <button string="Cancel" special="cancel" class="btn-secondary"/>
+                </footer>
+            </form>
+        </field>
+    </record>
+
+    <record id="action_report_purchase_retentions_wizard_wizard" model="ir.actions.act_window">
+        <field name="name">Compras y Retenciones</field>
+        <field name="res_model">report.purchase.retentions.wizard</field>
+        <field name="view_mode">form</field>
+        <field name="target">new</field>
+    </record>
+
+</odoo>
