+++ l10n_ec_reports_penta/wizard/report_sales_a1_wizard_views.xml
@@ -1,0 +1,28 @@
+<odoo>
+
+    <record id="view_report_sales_a1_wizard_form" model="ir.ui.view">
+        <field name="name">report.sales.a1.wizard.form</field>
+        <field name="model">report.sales.a1.wizard</field>
+        <field name="arch" type="xml">
+            <form string="Reporte Ventas (A1)">
+                <group>
+                    <field name="date_start"/>
+                    <field name="date_end"/>
+                    <field name="document_type"/>
+                </group>
+                <footer>
+                    <button name="print_report" type="object" string="Generate" class="btn-primary"/>
+                    <button string="Cancel" special="cancel" class="btn-secondary"/>
+                </footer>
+            </form>
+        </field>
+    </record>
+
+    <record id="action_report_sales_a1_wizard" model="ir.actions.act_window">
+        <field name="name">Reporte Ventas (A1)</field>
+        <field name="res_model">report.sales.a1.wizard</field>
+        <field name="view_mode">form</field>
+        <field name="target">new</field>
+    </record>
+
+</odoo>
