+++ l10n_ec_reports_penta/wizard/report_uafe_wizard_views.xml
@@ -1,0 +1,28 @@
+<odoo>
+
+    <record id="view_report_usafe_wizard_form" model="ir.ui.view">
+        <field name="name">report.uafe.wizard.form</field>
+        <field name="model">report.uafe.wizard</field>
+        <field name="arch" type="xml">
+            <form string="Report UAFE">
+                <group>
+                    <field name="year"/>
+                    <field name="month"/>
+                    <field name="domain_uafe"/>
+                </group>
+                <footer>
+                    <button name="print_report" type="object" string="Generate" class="btn-primary"/>
+                    <button string="Cancel" special="cancel" class="btn-secondary"/>
+                </footer>
+            </form>
+        </field>
+    </record>
+
+    <record id="action_report_uafe_wizard" model="ir.actions.act_window">
+        <field name="name">UAFE</field>
+        <field name="res_model">report.uafe.wizard</field>
+        <field name="view_mode">form</field>
+        <field name="target">new</field>
+    </record>
+
+</odoo>
